requirements.txt
binance-connector
tabulate
certifi